+++ packages/storybook-theme-tool/CHANGELOG.md
@@ -1,5 +1,23 @@
 # storybook-theme-tool
 
+## 2.8.12
+
+### Patch Changes
+
+- 2f7399d: publish
+- publish
+- Updated dependencies [2f7399d]
+- Updated dependencies
+  - @orderly.network/ui@2.8.13
+
+## 2.8.12-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/ui@2.8.13-alpha.0
+
 ## 2.8.11
 
 ### Patch Changes